test(fuzz): fuzz the session layer, not just the decoder
The fuzzer had one target. `wireformat` takes bytes and produces a
frame, which is a real thing to fuzz and also the smallest thing in the
system. The invariants that are actually worth guarding are about
*sequences* — a producer stops once its subscriber is gone, nothing
follows an ending, a unary call keeps working while subscriptions come
and go — and none of them can be expressed against a single decode.
Reaching them needs a server, a client, and a network that can be cut.

Turmoil already supplies all three deterministically, and
`turmoil_scenarios.rs` already had the runner and the oracle. What it
did not have was a way in other than counting seeds upwards. A seed
makes a finding reportable as one number, which is why the seed walk
stays; it also means the search is blind to whether the next seed
reaches anything the last one did not.

So the scenario core moves to `tests/scenario_body/mod.rs`, shared by
the seed walk, the new `scenarios` fuzz target and the regression
replay, for the reason `fuzz_body` already exists: a fuzzer and a replay
that have drifted apart both go green while the thing they guard has
moved. `Scenario::from_bytes` is the second way in — eight bytes of
network seed, then two bytes per action — so libfuzzer can keep the
inputs that reached new code instead of counting.

Both generators go through one `Action::from_pair`, so the two searches
cannot explore different spaces. That does change what a given seed
means; nothing is lost, because `tests/seeds/scenarios.txt` holds no
seeds yet.

Measured, not assumed. Two minutes on this machine: 13091 executions at
108/s, edge coverage 5582 to 5923, 434 new corpus units, no crashers and
no false leak reports. An hour-long night is roughly 390000 executions.

The workflow becomes a matrix over targets rather than a second copy of
two hundred lines, serial because both entries push the same corpus
branch and a race there throws away a night's work. Crash markers are
namespaced per target so two targets cannot collide on one issue. The
committed corpus is eight hand-written starting scenarios — read to the
end, read across a partition, drop without reading, subscribe while cut
— rather than the 406 the trial run produced; machine-found inputs
belong on the corpus branch, which is what it is for.
